+++ src/gameloop/valid_next_movement.c
@@ -3,21 +3,21 @@
 /*                                                        :::      ::::::::   */
 /*   valid_next_movement.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hdougoud <hdougoud@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jrandet <jrandet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/22 16:03:07 by jrandet           #+#    #+#             */
-/*   Updated: 2025/08/08 14:50:01 by hdougoud         ###   ########.fr       */
+/*   Updated: 2025/08/12 16:36:55 by jrandet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-void	valid_next_movement_up(t_main *main, double next_x, double next_y)
+void	valid_next_movement_up(t_main *main, float next_x, float next_y)
 {
-	double	current_x;
-	double	current_y;
-	double	valid_x;
-	double	valid_y;
+	float	current_x;
+	float	current_y;
+	float	valid_x;
+	float	valid_y;
 
 	current_x = main->player.x;
 	current_y = main->player.y;
@@ -35,12 +35,12 @@ void	valid_next_movement_up(t_main *main, double next_x, double next_y)
 		main->player.y = next_y;
 }
 
-void	valid_next_movement_down(t_main *main, double next_x, double next_y)
+void	valid_next_movement_down(t_main *main, float next_x, float next_y)
 {
-	double	current_x;
-	double	current_y;
-	double	valid_x;
-	double	valid_y;
+	float	current_x;
+	float	current_y;
+	float	valid_x;
+	float	valid_y;
 
 	current_x = main->player.x;
 	current_y = main->player.y;
@@ -58,12 +58,12 @@ void	valid_next_movement_down(t_main *main, double next_x, double next_y)
 		main->player.y = next_y;
 }
 
-void	valid_next_movement_right(t_main *main, double next_x, double next_y)
+void	valid_next_movement_right(t_main *main, float next_x, float next_y)
 {
-	double	current_x;
-	double	current_y;
-	double	valid_x;
-	double	valid_y;
+	float	current_x;
+	float	current_y;
+	float	valid_x;
+	float	valid_y;
 
 	current_x = main->player.x;
 	current_y = main->player.y;
@@ -81,12 +81,12 @@ void	valid_next_movement_right(t_main *main, double next_x, double next_y)
 		main->player.y = next_y;
 }
 
-void	valid_next_movement_left(t_main *main, double next_x, double next_y)
+void	valid_next_movement_left(t_main *main, float next_x, float next_y)
 {
-	double	current_x;
-	double	current_y;
-	double	valid_x;
-	double	valid_y;
+	float	current_x;
+	float	current_y;
+	float	valid_x;
+	float	valid_y;
 
 	current_x = main->player.x;
 	current_y = main->player.y;
