+++ src/utils/vector.c
@@ -6,13 +6,13 @@
 /*   By: jrandet <jrandet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/22 16:08:19 by jrandet           #+#    #+#             */
-/*   Updated: 2025/07/22 16:08:58 by jrandet          ###   ########.fr       */
+/*   Updated: 2025/08/12 16:36:55 by jrandet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-t_vec2	get_vector(double x1, double y1, double x2, double y2)
+t_vec2	get_vector(float x1, float y1, float x2, float y2)
 {
 	t_vec2	vector;
 
@@ -21,7 +21,7 @@ t_vec2	get_vector(double x1, double y1, double x2, double y2)
 	return (vector);
 }
 
-double	get_vector_norm(t_vec2 vec1, t_vec2 vec2)
+float	get_vector_norm(t_vec2 vec1, t_vec2 vec2)
 {
 	return (sqrt((pow(vec2.x - vec1.x, 2) + pow(vec2.y - vec1.y, 2))));
 }
